WebUI save-error — address review: share row-merge util, fall back to tab refresh on revert re-fetch failure, tidy tests
