+++ frontend/src/__tests__/containers/MasterWindow.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { act } from 'react-dom/test-utils';
-import { mount, shallow } from 'enzyme';
+import { mount } from 'enzyme';
 import nock from 'nock';
 import { Provider } from 'react-redux';
 import { applyMiddleware, combineReducers, createStore } from 'redux';
@@ -41,7 +41,6 @@ import viewHandler, {
 } from '../../reducers/viewHandler';
 import tables, {
   initialState as tablesHandlerState,
-  getTableId,
 } from '../../reducers/tables';
 
 import fixtures from '../../../test_setup/fixtures/master_window.json';
@@ -57,18 +56,6 @@ import topActionsFixtures
 import menuFixtures from '../../../test_setup/fixtures/master_window/menu.json';
 import userSessionData from '../../../test_setup/fixtures/user_session.json';
 import notificationsData from '../../../test_setup/fixtures/notifications.json';
-
-import MasterWindow from '../../components/app/MasterWindow';
-import { discardNewRequest, getRowsData } from '../../api';
-
-// closeModalCallback calls discardNewRequest (from ../../api) and awaits its
-// promise, then — for an already-persisted (numeric-id) abandoned row — re-fetches
-// that row via getRowsData and applies its reverted DB value to the parent grid.
-jest.mock('../../api', () => ({
-  __esModule: true,
-  discardNewRequest: jest.fn(() => Promise.resolve()),
-  getRowsData: jest.fn(() => Promise.resolve({ data: { result: [] } })),
-}));
 
 const middleware = [thunk, promiseMiddleware];
 const FIXTURES_PROPS = fixtures;
@@ -477,182 +464,4 @@ describe.skip('MasterWindowContainer', () => {
     });
 
   }, 20000);
-});
-
-describe('MasterWindow.closeModalCallback - abandon must not drop a persisted new-record row', () => {
-  const WINDOW_TYPE = '540321';
-  const DOCUMENT_ID = 'd1';
-  const TAB_ID = 't1';
-
-  // Minimal props so shallow() can construct the instance without render throwing.
-  const buildProps = (overrides = {}) => ({
-    modal: { visible: false },
-    master: {
-      docId: DOCUMENT_ID,
-      data: { DocumentNo: undefined },
-      layout: {},
-      includedTabsInfo: {},
-      hasComments: false,
-    },
-    breadcrumb: [],
-    rawModal: {},
-    pluginModal: {},
-    me: {},
-    overlay: { data: {}, visible: false },
-    params: { windowId: WINDOW_TYPE, docId: DOCUMENT_ID },
-    updateTabRowsData: jest.fn(),
-    ...overrides,
-  });
-
-  const getInstance = (props) => shallow(<MasterWindow {...props} />).instance();
-
-  // A realistic ?ids= re-fetch response: the persisted row, reverted server-side
-  // to its DB ValidFrom (a far-future year) after discardChanges — the abandoned
-  // in-memory colliding value (01/01/2015) is gone.
-  const PERSISTED_ROW_ID = '1234567';
-  const revertedRowsResponse = {
-    data: {
-      result: [
-        {
-          rowId: PERSISTED_ROW_ID,
-          fieldsByName: {
-            ValidFrom: { field: 'ValidFrom', value: '2053-11-30', widgetType: 'Date' },
-          },
-        },
-      ],
-    },
-  };
-
-  beforeEach(() => {
-    discardNewRequest.mockClear();
-    discardNewRequest.mockResolvedValue();
-    getRowsData.mockClear();
-    getRowsData.mockResolvedValue(revertedRowsResponse);
-  });
-
-  // Case A - the bug: a modal auto-saved (rowId became a real numeric id) but
-  // saveStatus stayed falsy. Abandon must NOT remove the now-persisted row.
-  it('does not remove a persisted (numeric rowId) row when saveStatus is falsy', async () => {
-    const props = buildProps();
-    const instance = getInstance(props);
-
-    await instance.closeModalCallback({
-      isNew: true,
-      windowType: WINDOW_TYPE,
-      documentId: DOCUMENT_ID,
-      tabId: TAB_ID,
-      rowId: PERSISTED_ROW_ID,
-      saveStatus: false,
-    });
-
-    const removedPersistedRow = props.updateTabRowsData.mock.calls.some(
-      ([, payload]) =>
-        payload && payload.removed && payload.removed[PERSISTED_ROW_ID]
-    );
-    expect(removedPersistedRow).toBe(false);
-  });
-
-  // Case B - regression guard: a genuinely unsaved new row (rowId === 'NEW')
-  // must still be discarded from the parent grid.
-  it('removes an unsaved new row (rowId === NEW) when saveStatus is falsy', async () => {
-    const props = buildProps();
-    const instance = getInstance(props);
-
-    await instance.closeModalCallback({
-      isNew: true,
-      windowType: WINDOW_TYPE,
-      documentId: DOCUMENT_ID,
-      tabId: TAB_ID,
-      rowId: 'NEW',
-      saveStatus: false,
-    });
-
-    const expectedTableId = getTableId({
-      windowId: WINDOW_TYPE,
-      docId: DOCUMENT_ID,
-      tabId: TAB_ID,
-    });
-    expect(props.updateTabRowsData).toHaveBeenCalledWith(expectedTableId, {
-      removed: { NEW: true },
-    });
-  });
-
-  // Case C - the fix: abandoning an already-persisted (numeric rowId) row with a
-  // falsy saveStatus must re-fetch the row and UPDATE (not remove) it in the grid,
-  // so the retained row shows its reverted DB value without a browser reload.
-  it('re-fetches and UPDATES a persisted row with its reverted DB value on abandon', async () => {
-    const props = buildProps();
-    const instance = getInstance(props);
-
-    await instance.closeModalCallback({
-      isNew: true,
-      windowType: WINDOW_TYPE,
-      documentId: DOCUMENT_ID,
-      tabId: TAB_ID,
-      rowId: PERSISTED_ROW_ID,
-      saveStatus: false,
-    });
-
-    const expectedTableId = getTableId({
-      windowId: WINDOW_TYPE,
-      docId: DOCUMENT_ID,
-      tabId: TAB_ID,
-    });
-
-    // the reverted DB row was re-fetched for exactly this persisted row
-    expect(getRowsData).toHaveBeenCalledWith(
-      expect.objectContaining({
-        entity: 'window',
-        docType: WINDOW_TYPE,
-        docId: DOCUMENT_ID,
-        tabId: TAB_ID,
-        rows: [PERSISTED_ROW_ID],
-      })
-    );
-
-    // it was applied as an UPDATE (changed), never a removal
-    const updateCall = props.updateTabRowsData.mock.calls.find(
-      ([, payload]) => payload && payload.changed && payload.changed[PERSISTED_ROW_ID]
-    );
-    expect(updateCall).toBeDefined();
-    expect(updateCall[0]).toBe(expectedTableId);
-    expect(updateCall[1].changed[PERSISTED_ROW_ID].fieldsByName.ValidFrom.value).toBe(
-      '2053-11-30'
-    );
-
-    const removedAny = props.updateTabRowsData.mock.calls.some(
-      ([, payload]) => payload && payload.removed
-    );
-    expect(removedAny).toBe(false);
-  });
-
-  // Case D - reconcile the deleted-row edge: if the persisted row was removed
-  // server-side between the abandon and the re-fetch, the ?ids= GET reports it in
-  // missingIds → the grid row must be removed (not left stranded).
-  it('removes a persisted row from the grid when the re-fetch reports it missing', async () => {
-    getRowsData.mockResolvedValue({
-      data: { result: [], missingIds: [PERSISTED_ROW_ID] },
-    });
-
-    const props = buildProps();
-    const instance = getInstance(props);
-
-    await instance.closeModalCallback({
-      isNew: true,
-      windowType: WINDOW_TYPE,
-      documentId: DOCUMENT_ID,
-      tabId: TAB_ID,
-      rowId: PERSISTED_ROW_ID,
-      saveStatus: false,
-    });
-
-    const expectedTableId = getTableId({
-      windowId: WINDOW_TYPE,
-      docId: DOCUMENT_ID,
-      tabId: TAB_ID,
-    });
-    expect(props.updateTabRowsData).toHaveBeenCalledWith(expectedTableId, {
-      removed: { [PERSISTED_ROW_ID]: true },
-    });
-  });
 });
